+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/message/attachment/AttachmentUploader.kt
@@ -3,11 +3,9 @@ package io.getstream.chat.android.offline.message.attachment
 import android.webkit.MimeTypeMap
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.call.await
-import io.getstream.chat.android.client.extensions.uploadId
 import io.getstream.chat.android.client.models.Attachment
-import io.getstream.chat.android.client.uploader.ProgressTrackerFactory
 import io.getstream.chat.android.client.uploader.StreamCdnImageMimeTypes
-import io.getstream.chat.android.client.uploader.toProgressCallback
+import io.getstream.chat.android.client.utils.ProgressCallback
 import io.getstream.chat.android.client.utils.Result
 import java.io.File
 
@@ -19,6 +17,7 @@ internal class AttachmentUploader(
         channelType: String,
         channelId: String,
         attachment: Attachment,
+        progressCallback: ProgressCallback? = null,
     ): Result<Attachment> {
         val file = checkNotNull(attachment.upload) { "An attachment needs to have a non null attachment.upload value" }
 
@@ -26,19 +25,11 @@ internal class AttachmentUploader(
             ?: attachment.mimeType
         val attachmentType = mimeType.toAttachmentType()
 
-        val progressTracker = attachment.uploadId?.let {
-            ProgressTrackerFactory.getOrCreate(it).apply {
-                maxValue = file.length()
-            }
-        }
-        val progressCallback = progressTracker?.toProgressCallback()
-
         val result = if (attachmentType == AttachmentType.IMAGE) {
             client.sendImage(channelType, channelId, file, progressCallback).await()
         } else {
             client.sendFile(channelType, channelId, file, progressCallback).await()
         }
-
         return if (result.isSuccess) {
             val augmentedAttachment = attachment.augmentAttachmentOnSuccess(
                 file = file,
@@ -46,10 +37,12 @@ internal class AttachmentUploader(
                 attachmentType = attachmentType,
                 url = result.data()
             )
-            progressTracker?.setComplete(true)
+            augmentedAttachment.uploadState = Attachment.UploadState.Success
+            progressCallback?.onSuccess(augmentedAttachment.url)
             Result(augmentedAttachment)
         } else {
-            progressTracker?.setComplete(false)
+            attachment.uploadState = Attachment.UploadState.Failed(result.error())
+            progressCallback?.onError(result.error())
             Result(result.error())
         }
     }