+++ CHANGELOG.md
@@ -9,8 +9,11 @@
 ### ✅ Added
 
 ### ⚠️ Changed
+- 🚨 Breaking change: `ProgressCallback` is not invoked on main thread anymore. So make sure to handle it if you were previously using this callback to update the UI directly.
+- Attachment#uploadState is now updated in real-time during uploads.
 
 ### ❌ Removed
+- Removed `ProgressTrackerFactory` and `ProgressTracker` in favour of new progress tracking implementation.
 
 ## stream-chat-android
 ### 🐞 Fixed
@@ -38,9 +41,12 @@
 - `ChatClient` now defaults to using the `https://chat.stream-io-api.com` base URL, using [Stream's Edge API Infrastructure](https://getstream.io/blog/chat-edge-infrastructure/) instead of connecting to a region-specific API. If you're not on a dedicated chat infrastructure, remove any region-specific base URL settings from the `ChatClient.Builder` to use Edge instead.
 
 ### ✅ Added
+- 🚨 Breaking change: A new `Idle` state is added to `Attachment.UploadState`.
+- Added a new callback function `onProgress(bytesUploaded: Long, totalLength: Long)` in `ProgressCallback`.
 - Added the possibility to add your own instance of OkHttpClient with `ChatClient.okHttpClient`.
 
 ### ⚠️ Changed
+- 🚨 Breaking change: `Attachment.UploadState.InProgress` now is data class having two fields, `bytesUploaded: Long` and `totalBytes: Long` instead of object.
 - Deprecated the `ChatClient.Builder#cdnUrl` method. To customize file uploads, set a custom `FileUploader` implementation instead. More info in the documentation: [Using Your Own CDN](https://getstream.io/chat/docs/android/file_uploads/?language=kotlin#using-your-own-cdn).
 
 ### ❌ Removed